Change expected result type to pass on 32-bit
Origin: upstream, https://github.com/scikit-learn/scikit-learn/issues/18084

Gbp-Pq: Name Change-expected-result-type-to-pass-on-32-bit.patch

+++ sklearn/datasets/tests/test_base.py
@@ -234,12 +234,12 @@ def test_load_breast_cancer():
 
 
 @pytest.mark.parametrize("loader_func, data_dtype, target_dtype", [
-    (load_breast_cancer, np.float64, np.int64),
+    (load_breast_cancer, np.float64, int),
     (load_diabetes, np.float64, np.float64),
-    (load_digits, np.float64, np.int64),
-    (load_iris, np.float64, np.int64),
+    (load_digits, np.float64, int),
+    (load_iris, np.float64, int),
     (load_linnerud, np.float64, np.float64),
-    (load_wine, np.float64, np.int64),
+    (load_wine, np.float64, int),
 ])
 def test_toy_dataset_as_frame(loader_func, data_dtype, target_dtype):
     default_result = loader_func()
